requirements.txt
# requirements.txt  (patched)

apache-airflow==2.9.0
feedparser
sqlalchemy
psycopg[binary]
pydantic

# --- added in patch sequence 1.1 ---
cryptography>=42.0.0      # used by bootstrap.sh for Fernet key
alembic                   # DB migrations (Step 3)
pytest                    # unit & smoke tests (Step 5)
pytest-asyncio            # future async scrapers
python-dotenv             # optional .env parsing in local scripts

# --- added in patch sequence 1.2 ---
praw                     # Reddit API client
telethon                # Telegram API client
gnewsclient            # Google News API client
scrapy                 # web scraping framework

# --- added in patch sequence 1.3 ---
beautifulsoup4          # HTML parsing for web scraping
requests                # HTTP requests for web scraping